fix(generic): update cloudflare challenge detection
