+++ src/generic/network.ts
@@ -1,7 +1,12 @@
 /* SPDX-License-Identifier: GPL-3.0-or-later */
 /* Copyright © 2026 Inkdex */
 
-import { PaperbackInterceptor, type Request, type Response } from "@paperback/types";
+import {
+  CloudflareError,
+  PaperbackInterceptor,
+  type Request,
+  type Response,
+} from "@paperback/types";
 import { MadaraGeneric } from "./main";
 
 export class MadaraInterceptor extends PaperbackInterceptor {
@@ -70,6 +75,22 @@ export class MadaraInterceptor extends PaperbackInterceptor {
     response: Response,
     data: ArrayBuffer,
   ): Promise<ArrayBuffer> {
+    const cfMitigated = response.headers?.["cf-mitigated"];
+    if (cfMitigated === "challenge") {
+      throw new CloudflareError(
+        {
+          url: this.source.bypassPage ? this.source.bypassPage : this.source.domain,
+          method: "GET",
+          headers: {
+            referer: `${this.source.domain}/`,
+            origin: `${this.source.domain}/`,
+            "user-agent": await Application.getDefaultUserAgent(),
+          },
+        },
+        "Cloudflare detected!\nPlease do the Cloudflare bypass to continue!",
+      );
+    }
+
     return data;
   }
 }
